# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hybrid firewall in Go using public blocklists",
        "title": "Meds: net healing",
        "contact": {
            "name": "cnaize",
            "url": "https://github.com/cnaize"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v1.2.0"
    },
    "paths": {
...
        "/v1/subnets/allowlist": {
            "get": {
                "description": "get all allowed subnets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Get allowed subnets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetSubnetsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert subnets to allowlist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Upsert allowed subnets",
                "parameters": [
                    {
                        "description": "subnets to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertSubnetsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove subnets from allowlist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Remove allowed subnets",
                "parameters": [
                    {
                        "description": "subnets to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveSubnetsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/subnets/allowlist/{ip}": {
            "get": {
                "description": "check if an ip address is allowed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Check allowed ip address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip address to check",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckSubnetResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/v1/subnets/blocklist/exclude": {
            "get": {
                "description": "get all excluded subnets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Get excluded subnets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetSubnetsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert subnets to excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Upsert excluded subnets",
                "parameters": [
                    {
                        "description": "subnets to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertSubnetsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove subnets from excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Remove excluded subnets",
                "parameters": [
                    {
                        "description": "subnets to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveSubnetsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/subnets/blocklist/exclude/{ip}": {
            "get": {
                "description": "check if an ip address is excluded",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Check excluded ip address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip address to check",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckSubnetResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/v1/subnets/blocklist/include": {
            "get": {
                "description": "get all included subnets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Get included subnets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetSubnetsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert subnets to includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Upsert included subnets",
                "parameters": [
                    {
                        "description": "subnets to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertSubnetsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove subnets from includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Remove included subnets",
                "parameters": [
                    {
                        "description": "subnets to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveSubnetsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/subnets/blocklist/include/{ip}": {
            "get": {
                "description": "check if an ip address is included",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subnets"
                ],
                "summary": "Check included ip address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip address to check",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckSubnetResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.CheckSubnetResp": {
            "type": "object",
            "properties": {
                "found": {
                    "type": "boolean"
                }
            }
        },
...
        "api.GetSubnetsResp": {
            "type": "object",
            "properties": {
                "subnets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "100.100.100.100/32",
                        "200.200.200.0/24"
                    ]
                }
            }
        },
...
        "api.RemoveSubnetsReq": {
            "type": "object",
            "properties": {
                "subnets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "100.100.100.100",
                        "200.200.200.0/24"
                    ]
                }
            }
        },
...
        "api.UpsertSubnetsReq": {
            "type": "object",
            "properties": {
                "subnets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "100.100.100.100",
                        "200.200.200.0/24"
                    ]
                }
            }
        }
    }
}